extract reusable atproto helpers; add docker-compose
- list-records.ts: pull paginateListRecords + rkeyFromCollectionAtUri
  out of standard-site-verify-backfill so other PDS crawls (next: the
  at.fund backfill) reuse one paginated listRecords loop.
- bluesky-public-profile: add fetchBlueskyPublicProfilesBatch using
  app.bsky.actor.getProfiles (max 25 actors per call), returning a
  Map<did, profile|null> in the same shape as the per-request memo the
  listing-mentions loader uses.
- docker-compose.yml gives local dev a one-command Postgres (pgvector/pg17).
